requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.19.0
pandas==1.0.5
scikit_learn==0.22.1
Flask-Cors==3.0.8
Flask-Compress==1.5.0
Unidecode==1.1.1
PyJWT==1.7.1


